Test fixes after refactoring
