+++ src/lms-ds-loader/tests/test_file_processor.py
@@ -6,7 +6,7 @@
 import pytest
 from unittest.mock import Mock
 
-from edfi_lms_ds_loader.constants import Constants
+from edfi_lms_ds_loader.helpers.constants import Table, Columns
 from edfi_lms_ds_loader.csv_to_sql import CsvToSql
 from edfi_lms_ds_loader.file_processor import FileProcessor
 from edfi_lms_ds_loader.lms_filesystem_provider import LmsFilesystemProvider
@@ -43,10 +43,10 @@ class Test_FileProcessor:
                 # Assert
                 assert file_processor.db_operations_adapter == db_operations_adapter
                 mock_csv_to_sql.assert_called_with(
-                    file_1, Constants.Table.USER, Constants.Columns.USER
+                    file_1, Table.USER, Columns.USER
                 )
                 mock_csv_to_sql.assert_called_with(
-                    file_2, Constants.Table.USER, Constants.Columns.USER
+                    file_2, Table.USER, Columns.USER
                 )
 
             def test_given_there_are_no_user_files(self, mocker):
